requirements.txt
langchain==0.0.305
langsmith==0.0.41
pandas==2.0.3
requests==2.31.0
openai==0.27.9
 google-api-python-client==2.105.0
 google-auth-httplib2==0.1.1
 google-auth-oauthlib==1.1.0
